Refactor and update report services
Refactored report services by removing and re-adding `OrderReportService` and `CustomerReportService` with updated namespaces and implementations. Removed `CustomerReportService.cs` and `OrderReportService.cs` files, and re-added them in new namespaces with enhanced methods and dependencies. Updated `ProductsReportService` namespace and removed `RegisterReports` method from `StartupApplication.cs`.
