+++ src/Business/Grand.Business.Customers/Startup/StartupApplication.cs
@@ -1,5 +1,6 @@
 ﻿using Grand.Business.Core.Interfaces.Customers;
 using Grand.Business.Core.Interfaces.ExportImport;
+using Grand.Business.Core.Interfaces.System.Reports;
 using Grand.Business.Customers.Services;
 using Grand.Business.Customers.Services.ExportImport;
 using Grand.Domain.Common;
@@ -35,6 +36,7 @@ public class StartupApplication : IStartupApplication
         serviceCollection.AddScoped<ICustomerNoteService, CustomerNoteService>();
         serviceCollection.AddScoped<ICustomerHistoryPasswordService, CustomerHistoryPasswordService>();
         serviceCollection.AddScoped<ICustomerManagerService, CustomerManagerService>();
+        serviceCollection.AddScoped<ICustomerReportService, CustomerReportService>();
         serviceCollection.AddScoped<ISalesEmployeeService, SalesEmployeeService>();
         serviceCollection.AddScoped<IUserApiService, UserApiService>();
         serviceCollection.AddScoped<IAffiliateService, AffiliateService>();
